Added a method that looks at the percent confirmed-case growth and retrieves all instances of provinces/states whose growth rate is above a user-provided threshold. It is called spikeAtTarget, and it takes 5 parameters: a dataframe representing the data observed; a String representing the target column to observe; a String value equal tot the field value of the target column to observe; a date that you want to 'query on' (i.e. how many confirmed cases after USA Valentine's Day (2/14/xx)?); a range of days to observe the 'spike'; a percentage as a decimal representing a threshold for distinguishing spikes.
By Default, the incubation period of COVID-19 is approximately 10-14 days, so we say that, the initial date of a spike is 14 days after your target day provided, and the date range provided is the number of days after your initial date plus 14. As listed above, suppose you want to know who got COVID from Valentine's Day. You would give 2/14/20 (or 21), and the program would look at days starting with February 28. If you say the range is 7 days, then from February 28th to March 6th is the time scale for the percent growth in confirmed cases. Each province/state in the region/country has a corresponding growth rate, if such a separation is given.